--- a/gradle.properties
+++ b/gradle.properties
@@ -4,3 +4,5 @@ version=0.12.0-SNAPSHOT
 # information required for publishing artifacts:
 edcScmConnection=scm:git:git@github.com:eclipse-edc/IdentityHub.git
 edcScmUrl=https://github.com/eclipse-edc/IdentityHub.git
+
+org.gradle.caching=true